完成

+++ src/main/java/org/spring/springboot/models/DetectionRequest.java
@@ -8,5 +8,5 @@ public class DetectionRequest {
     private Integer weight;
     private Integer waist;
     private Integer bust;
-    private Integer blood;
+    private Float blood;
 }
